rag-backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
together
sentence-transformers
numpy
openai
PyPDF2
python-docx
